+++ uxnandesktop/src/lib/components/ui/command/command-input.svelte
@@ -25,7 +25,7 @@
 			{...restProps}
 		>
 			{#snippet child({ props })}
-				<InputGroup.Input {...props} bind:value bind:ref />
+				<InputGroup.Input density="compact" {...props} bind:value bind:ref />
 			{/snippet}
 		</CommandPrimitive.Input>
 		<InputGroup.Addon>