+++ uxnandesktop/src/lib/components/ui/command/command.svelte
@@ -20,6 +20,6 @@
 	bind:value
 	bind:ref
 	data-slot="command"
-	class={cn("bg-popover text-popover-foreground rounded-xl! p-1 flex size-full flex-col overflow-hidden", className)}
+	class={cn("bg-popover text-popover-foreground rounded-xl! p-1 flex size-full max-w-[calc(100vw-1rem)] flex-col overflow-hidden", className)}
 	{...restProps}
 />