+++ uxnandesktop/src/lib/components/ui/tooltip/tooltip-provider.svelte
@@ -1,7 +1,21 @@
 <script lang="ts">
 	import { Tooltip as TooltipPrimitive } from "bits-ui";
 
-	let { delayDuration = 0, ...restProps }: TooltipPrimitive.ProviderProps = $props();
+	let {
+		delayDuration = 700,
+		skipDelayDuration = 0,
+		disableHoverableContent = true,
+		disableCloseOnTriggerClick = false,
+		ignoreNonKeyboardFocus = true,
+		...restProps
+	}: TooltipPrimitive.ProviderProps = $props();
 </script>
 
-<TooltipPrimitive.Provider {delayDuration} {...restProps} />
+<TooltipPrimitive.Provider
+	{delayDuration}
+	{skipDelayDuration}
+	{disableHoverableContent}
+	{disableCloseOnTriggerClick}
+	{ignoreNonKeyboardFocus}
+	{...restProps}
+/>